# fix(autoloading): Add missing autoloaders for shipped apps