--- a/apps/dashboard/composer/composer.json
+++ b/apps/dashboard/composer/composer.json
@@ -1,0 +1,13 @@
+{
+    "config" : {
+        "vendor-dir": ".",
+        "optimize-autoloader": true,
+        "classmap-authoritative": true,
+        "autoloader-suffix": "Dashboard"
+    },
+    "autoload" : {
+        "psr-4": {
+            "OCA\\Dashboard\\": "../lib/"
+        }
+    }
+}