--- a/apps/dashboard/composer/composer/installed.json
+++ b/apps/dashboard/composer/composer/installed.json
@@ -1,0 +1,5 @@
+{
+    "packages": [],
+    "dev": true,
+    "dev-package-names": []
+}