--- a/apps/files_external/composer/composer/installed.json
+++ b/apps/files_external/composer/composer/installed.json
@@ -1,0 +1,5 @@
+{
+    "packages": [],
+    "dev": true,
+    "dev-package-names": []
+}