--- a/apps/theming/composer/composer.json
+++ b/apps/theming/composer/composer.json
@@ -1,0 +1,13 @@
+{
+    "config" : {
+        "vendor-dir": ".",
+        "optimize-autoloader": true,
+        "classmap-authoritative": true,
+        "autoloader-suffix": "Theming"
+    },
+    "autoload" : {
+        "psr-4": {
+            "OCA\\Theming\\": "../lib/"
+        }
+    }
+}