--- a/apps/weather_status/composer/composer/installed.json
+++ b/apps/weather_status/composer/composer/installed.json
@@ -1,0 +1,5 @@
+{
+    "packages": [],
+    "dev": true,
+    "dev-package-names": []
+}