… apps/dashboard/composer/autoload.php
… apps/dashboard/composer/composer.lock
… apps/dashboard/composer/composer/InstalledVersions.php
… apps/dashboard/composer/composer/autoload_classmap.php
… apps/dashboard/composer/composer/autoload_namespaces.php
… apps/dashboard/composer/composer/autoload_psr4.php
… apps/dashboard/composer/composer/autoload_real.php
… apps/dashboard/composer/composer/autoload_static.php
… apps/files_external/composer/autoload.php
… apps/files_external/composer/composer.lock
… apps/files_external/composer/composer/ClassLoader.php
… apps/files_external/composer/composer/InstalledVersions.php
… apps/files_external/composer/composer/autoload_classmap.php
… apps/files_external/composer/composer/autoload_namespaces.php
… apps/files_external/composer/composer/autoload_psr4.php
… apps/files_external/composer/composer/autoload_real.php
… apps/files_external/composer/composer/autoload_static.php
… apps/theming/composer/autoload.php
… apps/theming/composer/composer.lock
… apps/theming/composer/composer/ClassLoader.php
… apps/theming/composer/composer/InstalledVersions.php
… apps/theming/composer/composer/autoload_classmap.php
… apps/theming/composer/composer/autoload_namespaces.php
… apps/theming/composer/composer/autoload_psr4.php
… apps/theming/composer/composer/autoload_real.php
… apps/theming/composer/composer/autoload_static.php
… apps/weather_status/composer/autoload.php
… apps/weather_status/composer/composer.lock
… apps/weather_status/composer/composer/ClassLoader.php
… apps/weather_status/composer/composer/InstalledVersions.php