@ChristophWurst 2023-01-12
> any reason why our build autoloader script did not cover those ? otherwise CI would have caught this discrepancy, would be good to add a step there for that

``if [[ -e ${app}/composer/composer.json ]]; then``

Only apps that have the composer file are checked