@ChristophWurst 2023-01-12
> otherwise CI would have caught this discrepancy, would be good to add a step there for that

Added