@ChristophWurst 2023-01-12
> files_external seems to have some issues still according to CI:
> 
> Migration step 'OCA\Files_External\Migration\Version1011Date20200630192246' is unknown [15](https://github.com/nextcloud/server/actions/runs/3900761480/jobs/6661803873#step:5:16)

Typo in the PSR4 spec